+++ client/src/components/streamPlayer.tsx
@@ -1,80 +1,31 @@
 import {
   Call,
   ParticipantView,
+  StatCardExplanation,
   StreamCall,
   StreamVideo,
   StreamVideoClient,
-  useCallStateHooks,
   User,
+  useCallStateHooks,
 } from "@stream-io/video-react-sdk";
-
-import { useEffect, useState } from "react";
-
 import { logError, logInfo } from "../lib/logger";
+import { useCallback, useEffect, useState } from "react";
+
+import { getStreamAdminCredentials } from "../lib/server-api";
+import { useAdminStore } from "../stores/adminStore";
 
 const user: User = { type: "anonymous" };
 
 interface VideoPlayerProps {
   muted: boolean;
-  streamCallId: string;
   fullscreen: boolean;
 }
 
 const StreamPlayer: React.FunctionComponent<VideoPlayerProps> = ({
   muted,
-  streamCallId,
   fullscreen,
 }) => {
-  const [state, setState] = useState<{
-    client: StreamVideoClient;
-    call: Call;
-  }>();
-
-  useEffect(() => {
-    const myClient = new StreamVideoClient({
-      apiKey: process.env.NEXT_PUBLIC_STREAM_API_KEY!,
-      user,
-    });
-    const myCall = myClient.call("livestream", streamCallId);
-
-    // myCall
-    //   .get()
-    //   .then(() => {
-    //     logInfo("Setting state for call ID " + streamCallId);
-    setState({ client: myClient, call: myCall });
-    // })
-    // .catch((e) => logError("Failed to retrieve call details", e));
-  }, [streamCallId]);
-
-  useEffect(() => {
-    if (!state) {
-      return;
-    }
-
-    state.call
-      .join()
-      .then(() => {})
-      .catch((e) => {
-        logError("Failed to join call", e);
-      });
-
-    return () => {
-      state.call.leave().catch((e) => {
-        logError("Failed to leave call", e);
-      });
-      setState(undefined);
-    };
-  }, [state]);
-
-  if (!state || !state.client || !state.call) return null;
-
-  return (
-    <StreamVideo client={state.client}>
-      <StreamCall call={state.call}>
-        <LivestreamView muted={muted} fullscreen={fullscreen} />
-      </StreamCall>
-    </StreamVideo>
-  );
+  return <LivestreamView muted={muted} fullscreen={fullscreen} />;
 };
 
 const LivestreamView = ({
@@ -95,17 +46,25 @@ const LivestreamView = ({
 
   console.log({ isLive, participants, liveParticipants });
 
-  return isLive && liveParticipants.length > 0 ? (
-    <ParticipantView
-      participant={liveParticipants[0]}
-      muteAudio={muted}
-      className={`noEvents stream-player ${fullscreen ? "videoAspectElement" : "streamPreviewAspectElement"}`}
-    />
-  ) : (
-    <div className="center:absolute whiteFill border padded:s1 highestLayer">
-      no video yet
-    </div>
-  );
+  if (isLive && liveParticipants.length > 0) {
+    console.log("YES VIDEOa");
+    return (
+      <>
+        <ParticipantView
+          participant={liveParticipants[0]}
+          muteAudio={muted}
+          className={`noEvents stream-player ${fullscreen ? "videoAspectElement" : "streamPreviewAspectElement"}`}
+        />
+      </>
+    );
+  } else {
+    console.log("NO VIDEO");
+    return (
+      <div className="center:absolute whiteFill border padded:s1 highestLayer">
+        no video yet
+      </div>
+    );
+  }
 };
 
 export default StreamPlayer;
